// flask-q4-recommendations-demo/new-q4-app/src/API.ts
export enum AnnotationTaskStatus {
  TODO = "TODO",
  FAILED = "FAILED",
  DONE = "DONE",
}
// ...
export type AnnotationTaskInstruction = {
  __typename: "AnnotationTaskInstruction",
  text: string,
  full_instructions_url?: string | null,
};

export enum AnnotationTaskType {
  NER = "NER",
  CLASSIFICATION = "CLASSIFICATION",
  ALLERGIES = "ALLERGIES",
}


export type AnnotationTag = {
  __typename: "AnnotationTag",
  id: string,
  name: string,
  description: string,
};

export type ClassificationAnnotationSample = {
  __typename: "ClassificationAnnotationSample",
  text: string,
  annotations:  Array<ClassificationAnnotation | null >,
};

export type ClassificationAnnotation = {
  __typename: "ClassificationAnnotation",
  text: string,
  tag?: AnnotationTag | null,
};

export type AnnotationTaskIssue = {
  __typename: "AnnotationTaskIssue",
  id: string,
  type: AnnotationTaskIssueType,
  reason?: AnnotationTaskIssueReason | null,
  notes: string,
  createdAt: string,
  updatedAt: string,
};

export type AnnotationTaskIssueType = {
  __typename: "AnnotationTaskIssueType",
  id: string,
  type: string,
  createdAt: string,
  updatedAt: string,
};

export type AnnotationTaskIssueReason = {
  __typename: "AnnotationTaskIssueReason",
  id: string,
  reason: string,
  createdAt: string,
  updatedAt: string,
};

export type AllergyAnnotationTask = {
  __typename: "AllergyAnnotationTask",
  id: string,
  project_id: string,
  instruction: AnnotationTaskInstruction,
  status: AnnotationTaskStatus,
  type: AnnotationTaskType,
  metadata?: string | null,
  annotator_id: string,
  tags:  Array<AnnotationTag >,
  sample: AllergyAnnotationSample,
  issue?: AnnotationTaskIssue | null,
  createdAt: string,
  updatedAt: string,
};

export type AllergyAnnotationSample = {
  __typename: "AllergyAnnotationSample",
  text: string,
  annotations?:  Array<AnnotationTagWithEvidence | null > | null,
};

export type AnnotationTagWithEvidence = {
  __typename: "AnnotationTagWithEvidence",
  id: string,
  tag?: AnnotationTag | null,
  evidence?: ModelEvidenceConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelEvidenceConnection = {
  __typename: "ModelEvidenceConnection",
  items?:  Array<Evidence | null > | null,
  nextToken?: string | null,
};

export type Evidence = {
  __typename: "Evidence",
  id: string,
  annotation_id: string,
  citation_snippet: string,
  last_updated: string,
  annotator_id: string,
  data_source: DataSource,
  title: string,
  url?: string | null,
  year?: string | null,
  doi?: string | null,
  author?: string | null,
  notes?: string | null,
  pages_used?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type IEvidence = {
  __typename: "IEvidence",
  id: string,
  title: string,
  url?: string | null,
  year?: string | null,
  citation_snippet?: string | null,
  doi?: string | null,
  author?: string | null,
  notes?: string | null,
  last_updated: string,
  annotator_id: string,
  data_source: DataSource,
};

export type DataSource = {
  __typename: "DataSource",
  id: string,
  name: string,
  description?: string | null,
  url?: string | null,
  type: DataSourceType,
  publisher?: string | null,
  volume_no?: string | null,
  issue_no?: string | null,
  edition?: string | null,
  year?: string | null,
  projects?: ModelProjectDataSourceConnection | null,
  createdAt: string,
  updatedAt: string,
};

export enum DataSourceType {
  Journal = "Journal",
  Book = "Book",
  Url = "Url",
}


export type ModelProjectDataSourceConnection = {
  __typename: "ModelProjectDataSourceConnection",
  items?:  Array<ProjectDataSource | null > | null,
  nextToken?: string | null,
};